BF: fix building debian packages on ubuntu
As discussed in the thread:

http://lists.alioth.debian.org/pipermail/neurodebian-users/2012-May/000411.html

nd_adddist fails to create cow images for debian versions when using it on
ubuntu.

The error is:

E: Release signed by unknown key (key id AED4B06F473041FA)

To fix this, we need to explicitly tell it to use the debian archive keyring
when building on ubuntu.